+++ packages/scheduler-sequelize/test/helpers_test.ts
@@ -1,7 +1,7 @@
 import moment, { Moment } from 'moment-timezone';
 import { expect } from 'chai';
 import sinon, { SinonSandbox, SinonFakeTimers } from 'sinon';
-import { computeNextRun, isHealthy } from '../src/helpers';
+import { computeNextRun, computeNextRuns, isHealthy } from '../src/helpers';
 
 describe('helpers', () => {
   let sandbox: SinonSandbox;
@@ -18,11 +18,20 @@ describe('helpers', () => {
     sandbox.restore();
   });
 
-  describe('compute next run at', () => {
+  describe('computeNextRun', () => {
     it('Calculates the next date correctly', () => {
       const every3Hours = 'FREQ=HOURLY;INTERVAL=4;BYMINUTE=0';
       const nextDate = moment(computeNextRun(every3Hours));
       expect(nextDate.diff(moment().tz('utc').minute(0), 'hours')).to.equal(3);
+    });
+
+    it('Calculates the next date correctly with DTSTART rule', () => {
+      const start = moment().tz('UTC').second(0).format('YYYYMMDDTHHmmss')
+      const every3Hours =
+        `DTSTART;TZID=UTC:${start}\nRRULE:FREQ=HOURLY;BYMINUTE=0;INTERVAL=3`;
+      const nextRun = moment(computeNextRun(every3Hours));
+      const nextDate = moment().tz('Australia/Sydney').add(3, 'hours').minute(0).second(0).millisecond(0);
+      expect(nextRun.toISOString()).to.equal(nextDate.toISOString());
     });
 
     // List of recurrence rules to test
@@ -132,6 +141,54 @@ describe('helpers', () => {
         parsed.minutes(),
         parsed.format('ZZ'),
       ]).to.eqls([17, 3, 2024, 12, 0, '+1000']);
+    });
+  });
+
+  describe('computeNextRuns', () => {
+    it('Calculates the next dates correctly', () => {
+      const every3Hours = 'FREQ=HOURLY;INTERVAL=4;BYMINUTE=0';
+      const [nextDates] = computeNextRuns(every3Hours);
+      const nextDate = moment(nextDates);
+      expect(nextDate.diff(moment().tz('utc').minute(0), 'hours')).to.equal(3);
+    });
+
+    it('Calculates the next dates correctly with DTSTART rule', () => {
+      const start = moment().tz('UTC').second(0).format('YYYYMMDDTHHmmss')
+      const every3Hours =
+        `DTSTART;TZID=UTC:${start}\nRRULE:FREQ=HOURLY;BYMINUTE=0;INTERVAL=3`;
+      const [nextRuns] = computeNextRuns(every3Hours);
+      const nextRun = moment(nextRuns);
+      const nextDate = moment().tz('Australia/Sydney').add(3, 'hours').minute(0).second(0).millisecond(0);
+      expect(nextRun.toISOString()).to.equal(nextDate.toISOString());
+    });
+
+    it('Returns correct number of next runs', () => {
+      const every3Hours = 'FREQ=HOURLY;INTERVAL=4;BYMINUTE=0';
+      const nextDates = computeNextRuns(every3Hours, { count: 10 });
+      expect(nextDates).to.have.length(10);
+    });
+
+    it('Returns correct number of next runs with DTSTART rule', () => {
+      const every3Hours =
+        'DTSTART;TZID=Australia/Sydney:20240120T040000\nRRULE:FREQ=HOURLY;BYMINUTE=0;INTERVAL=4';
+      const nextDates = computeNextRuns(every3Hours, { count: 10 });
+      expect(nextDates).to.have.length(10);
+    });
+
+    it('Lunartick rule and DTSTART rule should match', () => {
+      const lunartickRecurrence = 'FREQ=DAILY;INTERVAL=1;BYMINUTE=0;BYSECOND=0;';
+      const start = moment().tz('UTC').millisecond(0).format('YYYYMMDDTHHmmss')
+      const rruleRecurrence =
+        `DTSTART;TZID=UTC:${start}\nRRULE:FREQ=DAILY;BYMINUTE=0;BYSECOND=0;INTERVAL=1`;
+
+      const lunartickDates = computeNextRuns(lunartickRecurrence, {
+        count: 10,
+      });
+      const rruleDates = computeNextRuns(rruleRecurrence, {
+        count: 10,
+      });
+
+      expect(lunartickDates).to.deep.equal(rruleDates);
     });
   });